pyproject.toml
[project]
name = "llm-chunking"
version = "0.1.0"
description = "A hands-on learning project exploring LiteLLM, text chunking strategies, and vector databases"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chonkie[st]>=1.3.0",
    "isort>=6.0.1",
    "numpy>=2.3.3",
    "openai>=1.108.0",
    "openai-agents>=0.3.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "ruff>=0.13.1",
]
